automate-gateway/handler/iamv2beta: add CreateRule handler
This is so tedious. It's almost, but not exactly, the same code as that
written in authz-service, translating from its GRPC service to its
internal storage interface (and back).

One thing to note: we're using an extra ENUM value in our protobuf
definitions in the gateway API. It's a way to ensure that the enum is
explicitly provided: if it's left off, the request message with not have
the type that happens to be the first enum name, but an "unset" enum
that allows discrimination.  (Wrapping the enum could help, too, but
this is neat and tidy and straightforward.)

Signed-off-by: Stephan Renatus <srenatus@chef.io>